添加误差调整范围以及数据同步按钮，同步按钮功能尚未添加

+++ software/Firmware/APP/water_flow_time_control_sys.c
@@ -265,7 +265,7 @@ static uint8_t getFlow(uint16_t *Flow)
 }
 
 /* 返回允许的误差值 */
-float getTolerance(void)
+uint16_t getTolerance(void)
 {
     return memData.flowVsTime.tolerance;
 }
